oci8: Fix GetRowAssoc with ADODB_FETCH_ASSOC fetch mode
With Oracle, calls to GetRowAssoc(0) fail to return data in the
associative array (all fields are set to null), because the recordset's
field names are in upper case.

The revised GetAssocKeys() function takes the recordset's fetch mode
into consideration to build the bind property, ensuring that GetRowAssoc
is able to retrieve the results as expected.

Reference: http://phplens.com/lens/lensforum/msgs.php?id=17539